backend/requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
accelerate
safetensors
Pillow
opencv-python
numpy
torchvision
